fixed issues that came from merge.
